nyanya_backend/requirements.txt
# Core Django and API
Django==5.0.6
djangorestframework==3.15.2

# Database
psycopg[binary]==3.2.3

# API Documentation
drf-yasg==1.21.7

# CORS handling for frontend
django-cors-headers==4.4.0

# Environment variables
python-decouple==3.8

# Data processing & ML
pandas==2.2.3
scikit-learn==1.6.1
requests==2.32.3
